# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Book Quiz Lab — AI Material Generator
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface Book {
  id: number;
  title: string;
  description?: string;
  coverImageUrl?: string;
  createdAt: string;
}

export interface BookInput {
  title: string;
  description?: string;
  coverImageUrl?: string;
}

export interface BookUpdate {
  title?: string;
  description?: string;
  coverImageUrl?: string;
}

export interface MaterialSummary {
  id: number;
  bookId: number;
  bookTitle: string;
  chapterTitle: string;
  level: string;
  vocabCount: number;
  vocabQuizCount: number;
  readingQuizCount: number;
  createdAt: string;
}

export type GenerateInputLevel = typeof GenerateInputLevel[keyof typeof GenerateInputLevel];


export const GenerateInputLevel = {
  elementary4: 'elementary4',
  elementary5: 'elementary5',
  elementary6: 'elementary6',
  middle: 'middle',
} as const;

export interface GenerateInput {
  chapterTitle: string;
  level: GenerateInputLevel;
}

export interface GenerateResult {
  materialId: number;
  vocabListPdfBase64: string;
  vocabQuizPdfBase64: string;
  readingQuizPdfBase64: string;
  answerKeyPdfBase64: string;
}

export interface PdfBundle {
  vocabListPdfBase64: string;
  vocabQuizPdfBase64: string;
  readingQuizPdfBase64: string;
  answerKeyPdfBase64: string;
}

export interface UploadRequestBody {
  filename: string;
  contentType: string;
  folder?: string;
}

export interface UploadResponseData {
  uploadUrl: string;
  objectPath: string;
  publicUrl?: string;
}