# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Book Quiz Lab — AI Material Generator
servers:
  - url: /api
tags:
  - name: health
  - name: books
  - name: materials
  - name: storage
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List all books
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: createBook
      tags: [books]
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"

  /books/{bookId}:
    patch:
      operationId: updateBook
      tags: [books]
      summary: Update a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /books/{bookId}/materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List generated materials for a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaterialSummary"

  /books/{bookId}/generate:
    post:
      operationId: generateMaterials
      tags: [materials]
      summary: AI-generate all materials and save to archive
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInput"
      responses:
        "201":
          description: Generated materials with PDF data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResult"

  /materials/{materialId}:
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material set
      parameters:
        - name: materialId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /materials/{materialId}/pdf:
    post:
      operationId: downloadMaterialPdf
      tags: [materials]
      summary: Re-generate PDFs from archived material
      parameters:
        - name: materialId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfBundle"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned URL to upload a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadRequestBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponseData"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Book:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        coverImageUrl:
          type: string
        createdAt:
          type: string
          format: date-time

    BookInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        coverImageUrl:
          type: string

    BookUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        coverImageUrl:
          type: string

    MaterialSummary:
      type: object
      required: [id, bookId, bookTitle, chapterTitle, level, vocabCount, vocabQuizCount, readingQuizCount, createdAt]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        bookTitle:
          type: string
        chapterTitle:
          type: string
        level:
          type: string
        vocabCount:
          type: integer
        vocabQuizCount:
          type: integer
        readingQuizCount:
          type: integer
        createdAt:
          type: string
          format: date-time

    GenerateInput:
      type: object
      required: [chapterTitle, level]
      properties:
        chapterTitle:
          type: string
        level:
          type: string
          enum: [elementary4, elementary5, elementary6, middle]

    GenerateResult:
      type: object
      required: [materialId, vocabListPdfBase64, vocabQuizPdfBase64, readingQuizPdfBase64, answerKeyPdfBase64]
      properties:
        materialId:
          type: integer
        vocabListPdfBase64:
          type: string
        vocabQuizPdfBase64:
          type: string
        readingQuizPdfBase64:
          type: string
        answerKeyPdfBase64:
          type: string

    PdfBundle:
      type: object
      required: [vocabListPdfBase64, vocabQuizPdfBase64, readingQuizPdfBase64, answerKeyPdfBase64]
      properties:
        vocabListPdfBase64:
          type: string
        vocabQuizPdfBase64:
          type: string
        readingQuizPdfBase64:
          type: string
        answerKeyPdfBase64:
          type: string

    UploadRequestBody:
      type: object
      required: [filename, contentType]
      properties:
        filename:
          type: string
        contentType:
          type: string
        folder:
          type: string

    UploadResponseData:
      type: object
      required: [uploadUrl, objectPath]
      properties:
        uploadUrl:
          type: string
        objectPath:
          type: string
        publicUrl:
          type: string