# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Book Quiz Lab — AI Material Generator
 * OpenAPI spec version: 0.2.0
 */

export interface Book {
  id: number;
  title: string;
  description?: string;
  coverImageUrl?: string;
  createdAt: Date;
}