# lib/api-zod/src/generated/types/generateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Book Quiz Lab — AI Material Generator
 * OpenAPI spec version: 0.2.0
 */
import type { GenerateInputLevel } from './generateInputLevel';

export interface GenerateInput {
  chapterTitle: string;
  level: GenerateInputLevel;
}