# lib/api-zod/src/generated/types/materialSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Book Quiz Lab — AI Material Generator
 * OpenAPI spec version: 0.2.0
 */

export interface MaterialSummary {
  id: number;
  bookId: number;
  bookTitle: string;
  chapterTitle: string;
  level: string;
  vocabCount: number;
  vocabQuizCount: number;
  readingQuizCount: number;
  createdAt: Date;
}