# lib/api-zod/src/generated/types/pdfBundle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Book Quiz Lab — AI Material Generator
 * OpenAPI spec version: 0.2.0
 */

export interface PdfBundle {
  vocabListPdfBase64: string;
  vocabQuizPdfBase64: string;
  readingQuizPdfBase64: string;
  answerKeyPdfBase64: string;
}