# lib/api-zod/src/generated/types/uploadRequestBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Book Quiz Lab — AI Material Generator
 * OpenAPI spec version: 0.2.0
 */

export interface UploadRequestBody {
  filename: string;
  contentType: string;
  folder?: string;
}