# lib/api-zod/src/generated/types/uploadResponseData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Book Quiz Lab — AI Material Generator
 * OpenAPI spec version: 0.2.0
 */

export interface UploadResponseData {
  uploadUrl: string;
  objectPath: string;
  publicUrl?: string;
}